+++ scripts/tmp-sweep.sh
@@ -101,10 +101,10 @@ clodex-hooks-
 clodex-hs-
 clodex-impwire-
 clodex-impwire-home-
-clodex-injbit-
 clodex-impwire-reg-
 clodex-impwire-src-
 clodex-init-
+clodex-injbit-
 clodex-install-line-
 clodex-ipccache-
 clodex-jsonlw-